+++ NEWS.md
@@ -1,5 +1,7 @@
 # activatr (development version)
 
+* Switch dependency from `timetk` to `slider` to avoid unnecessary indirection (#3).
+
 # activatr 0.1.0
 
 * Initial release.